merge

+++ editor/Processes/RemoteProcessWorkerContainer.cs
@@ -11,32 +11,27 @@ namespace StorybrewEditor.Processes
 {
     public class RemoteProcessWorkerContainer : IDisposable
     {
-        private IChannel channel;
-        private Process process;
+        IChannel channel;
+        Process process;
 
         public RemoteProcessWorker Worker { get; private set; }
 
         public RemoteProcessWorkerContainer()
         {
-            var identifier = $"{Guid.NewGuid().ToString()}";
+            var identifier = $"{Guid.NewGuid()}";
             var workerUrl = $"ipc://sbrew-worker-{identifier}/worker";
 
-            channel = new IpcChannel(
-                new Hashtable()
-                {
-                    ["name"] = $"sbrew-{identifier}",
-                    ["portName"] = $"sbrew-{identifier}",
-                },
-                new BinaryClientFormatterSinkProvider(),
-                null
-            );
-            ChannelServices.RegisterChannel(channel, false);
+            channel = new IpcChannel(new Hashtable()
+            {
+                ["name"] = $"sbrew-{identifier}",
+                ["portName"] = $"sbrew-{identifier}"
+            }, new BinaryClientFormatterSinkProvider(), null);
 
+            ChannelServices.RegisterChannel(channel, false);
             startProcess(identifier);
             Worker = retrieveWorker(workerUrl);
         }
-
-        private void startProcess(string identifier)
+        void startProcess(string identifier)
         {
             var executablePath = Assembly.GetExecutingAssembly().Location;
             var workingDirectory = Path.GetDirectoryName(executablePath);
@@ -44,13 +39,12 @@ namespace StorybrewEditor.Processes
             {
                 StartInfo = new ProcessStartInfo(executablePath, $"worker \"{identifier}\"")
                 {
-                    WorkingDirectory = workingDirectory,
+                    WorkingDirectory = workingDirectory
                 },
             };
             process.Start();
         }
-
-        private RemoteProcessWorker retrieveWorker(string workerUrl)
+        RemoteProcessWorker retrieveWorker(string workerUrl)
         {
             while (true)
             {
@@ -71,11 +65,10 @@ namespace StorybrewEditor.Processes
 
         #region IDisposable Support
 
-        private bool disposedValue = false;
-
+        bool disposed = false;
         protected virtual void Dispose(bool disposing)
         {
-            if (!disposedValue)
+            if (!disposed)
             {
                 if (disposing)
                 {
@@ -87,21 +80,16 @@ namespace StorybrewEditor.Processes
                     {
                         Trace.WriteLine($"Failed to dispose the worker: {e}");
                     }
-                    if (!process.WaitForExit(3000))
-                        process.Kill();
+                    if (!process.WaitForExit(3000)) process.Kill();
                     ChannelServices.UnregisterChannel(channel);
                 }
                 Worker = null;
                 process = null;
                 channel = null;
-                disposedValue = true;
+                disposed = true;
             }
         }
-
-        public void Dispose()
-        {
-            Dispose(true);
-        }
+        public void Dispose() => Dispose(true);
 
         #endregion
     }
